rollback `SequentialIntChannel` renaming
