+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/specifications/SequentialIntChannel.kt
